Rename file of update-schedules query hook
